+++ test/vhpi/vhpi_test.c
@@ -65,6 +65,7 @@ static const vhpi_test_t tests[] = {
    { "issue1463", issue1463_startup },
    { "issue1473", issue1473_startup },
    { "issue1505", issue1505_startup },
+   { "issue1616", issue1616_startup },
    { NULL,        NULL },
 };