+++ test/vhpi/vhpi_test.h
@@ -77,5 +77,6 @@ void issue1428_startup(void);
 void issue1463_startup(void);
 void issue1473_startup(void);
 void issue1505_startup(void);
+void issue1616_startup(void);
 
 #endif  // _VHPI_TEST_H